+++ cmd.go
@@ -415,6 +415,10 @@ func displayNumberMenu(ctx context.Context, pkgS []string, dbExecutor db.Executo
 		return nil
 	}
 
+	if config.NewInstallEngine {
+		return syncInstall(ctx, config, cmdArgs, dbExecutor)
+	}
+
 	return install(ctx, arguments, dbExecutor, true)
 }
